requirement.txt
Django==4.2.0
Pillow==9.5.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7